requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==4.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
django-silk==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Faker==28.0.0
Flask==3.0.3
gprof2dot==2024.6.6
idna==3.8
importlib_metadata==8.0.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.12.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
toposort==1.10
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
